:art: Separate `log_binary` library
Problem:
- `log` and `log_binary` are two separate libraries but they share a directory.

Solution:
- Separate `log_binary` into its own directory.
