+++ include/log_binary/catalog/arguments.hpp
