+++ include/log_binary/catalog/mipi_messages.hpp
