+++ test/CMakeLists.txt
@@ -20,6 +20,7 @@ add_subdirectory(cib)
 add_subdirectory(flow)
 add_subdirectory(interrupt)
 add_subdirectory(log)
+add_subdirectory(log_binary)
 add_subdirectory(log_fmt)
 add_subdirectory(lookup)
 add_subdirectory(match)
