+++ test/log/log.cpp
@@ -1,5 +1,5 @@
-#include <log_fmt/logger.hpp>
 #include <log/log.hpp>
+#include <log_fmt/logger.hpp>
 
 #include <stdx/ct_string.hpp>
 #include <stdx/panic.hpp>
