+++ test/log_binary/catalog_app.cpp
