+++ test/log_binary/catalog_concurrency.hpp
