+++ test/log_binary/catalog_enums.hpp
