+++ test/log_binary/encoder.cpp
@@ -1,4 +1,4 @@
-#include <log/catalog/encoder.hpp>
+#include <log_binary/catalog/encoder.hpp>
 
 #include <stdx/concepts.hpp>
 #include <stdx/ct_format.hpp>
