+++ test/log_binary/stable_strings.json
